moved patch that removes invocation unlearn to transpiler on `CharacterBuildingManager.GrantFeatures`
